add stale guilds cleanup config setting

+++ bonkbot/bot.py
@@ -447,9 +447,14 @@ class BonkBot(discord.Client):
         self.__logger.debug(f"Guilds that are cached in database: {list(guilds_cached)}")
         self.__logger.debug(f"Difference: {stale_guild_ids}")
         
+        if not self.__config.clean_up_stale_guilds:
+            self.__logger.info(f"Skipping deletion of stale guilds as feature is disabled")
+            return
+        
         if len(stale_guild_ids) == 0:
             self.__logger.info(f"No stale guilds found, exiting job")
             return
+        
         self.__logger.info(f"Removing stale guilds from database: {stale_guild_ids}")
         self.__data_service.delete_guilds(stale_guild_ids)
